cleanup and addition of margin + collateral loan
