+++ token-lending/program/src/instruction.rs
@@ -172,28 +172,11 @@ pub enum LendingInstruction {
     MarginBorrowReserveLiquidity {
         // TODO: slippage constraint
         /// Amount whose usage depends on `amount_type`
-        amount: u64,
+        collateral_amount: u64,
+        /// amount
+        loan_amount: u64,
         /// Describe how the amount should be treated
         amount_type: BorrowAmountType,
-    },
-
-    /// Repay loaned tokens to a reserve. This loan was taken with no obligation, interest rates must be paid back at this step.
-    ///
-    ///   0. `[writable]` Source liquidity token account, minted by repay reserve liquidity mint
-    ///                     $authority can transfer $collateral_amount
-    ///   1. `[writable]` Destination collateral token account, minted by withdraw reserve collateral mint
-    ///   2. `[writable]` Repay reserve account.
-    ///   3. `[writable]` Repay reserve liquidity supply SPL Token account
-    ///   4. `[]` Withdraw reserve account.
-    ///   5. `[writable]` Withdraw reserve collateral supply SPL Token account
-    ///   9. `[]` Lending market account.
-    ///   10 `[]` Derived lending market authority.
-    ///   11 `[]` User transfer authority ($authority).
-    ///   12 `[]` Clock sysvar
-    ///   13 `[]` Token program id
-    MarginRepayReserveLiquidity {
-        /// Amount of loan to repay
-        liquidity_amount: u64,
     },
     /// Repay loaned tokens to a reserve and receive collateral tokens. The obligation balance
     /// will be recalculated for interest.
@@ -416,18 +399,16 @@ impl LendingInstruction {
                 buf.extend_from_slice(&amount_type.to_u8().unwrap().to_le_bytes());
             }
             Self::MarginBorrowReserveLiquidity {
-                amount,
+                collateral_amount,
+                loan_amount,
                 amount_type,
             } => {
                 buf.push(5);
-                buf.extend_from_slice(&amount.to_le_bytes());
-                // buf.extend_from_slice(&amount_type.to_u8().unwrap().to_le_bytes());
+                buf.extend_from_slice(&collateral_amount.to_le_bytes());
+                buf.extend_from_slice(&loan_amount.to_le_bytes());
+                buf.extend_from_slice(&amount_type.to_u8().unwrap().to_le_bytes());
             }
             Self::RepayReserveLiquidity { liquidity_amount } => {
-                buf.push(6);
-                buf.extend_from_slice(&liquidity_amount.to_le_bytes());
-            }
-            Self::MarginRepayReserveLiquidity { liquidity_amount } => {
                 buf.push(6);
                 buf.extend_from_slice(&liquidity_amount.to_le_bytes());
             }
@@ -682,7 +663,8 @@ pub fn borrow_reserve_liquidity(
 #[allow(clippy::too_many_arguments)]
 pub fn margin_borrow_reserve_liquidity(
     program_id: Pubkey,
-    amount: u64,
+    collateral_amount: u64,
+    loan_amount: u64,
     amount_type: BorrowAmountType,
     source_collateral_pubkey: Pubkey,
     destination_liquidity_pubkey: Pubkey,
@@ -726,7 +708,8 @@ pub fn margin_borrow_reserve_liquidity(
         program_id,
         accounts,
         data: LendingInstruction::MarginBorrowReserveLiquidity {
-            amount,
+            collateral_amount,
+            loan_amount,
             amount_type,
         }
         .pack(),
@@ -770,40 +753,6 @@ pub fn repay_reserve_liquidity(
             AccountMeta::new_readonly(spl_token::id(), false),
         ],
         data: LendingInstruction::RepayReserveLiquidity { liquidity_amount }.pack(),
-    }
-}
-
-/// Creates a `MarginRepayReserveLiquidity` instruction
-#[allow(clippy::too_many_arguments)]
-pub fn margin_repay_reserve_liquidity(
-    program_id: Pubkey,
-    liquidity_amount: u64,
-    source_liquidity_pubkey: Pubkey,
-    destination_collateral_pubkey: Pubkey,
-    repay_reserve_pubkey: Pubkey,
-    repay_reserve_liquidity_supply_pubkey: Pubkey,
-    withdraw_reserve_pubkey: Pubkey,
-    withdraw_reserve_collateral_supply_pubkey: Pubkey,
-    lending_market_pubkey: Pubkey,
-    lending_market_authority_pubkey: Pubkey,
-    user_transfer_authority_pubkey: Pubkey,
-) -> Instruction {
-    Instruction {
-        program_id,
-        accounts: vec![
-            AccountMeta::new(source_liquidity_pubkey, false),
-            AccountMeta::new(destination_collateral_pubkey, false),
-            AccountMeta::new(repay_reserve_pubkey, false),
-            AccountMeta::new(repay_reserve_liquidity_supply_pubkey, false),
-            AccountMeta::new_readonly(withdraw_reserve_pubkey, false),
-            AccountMeta::new(withdraw_reserve_collateral_supply_pubkey, false),
-            AccountMeta::new_readonly(lending_market_pubkey, false),
-            AccountMeta::new_readonly(lending_market_authority_pubkey, false),
-            AccountMeta::new_readonly(user_transfer_authority_pubkey, true),
-            AccountMeta::new_readonly(sysvar::clock::id(), false),
-            AccountMeta::new_readonly(spl_token::id(), false),
-        ],
-        data: LendingInstruction::MarginRepayReserveLiquidity { liquidity_amount }.pack(),
     }
 }
